+++ packages/design-tokens/src/system/transitions.ts
@@ -24,12 +24,21 @@
 
  */
 
-export interface Transitions {
-  none: number
-  rapid: number
-  quick: number
-  simple: number
-  moderate: number
-  complex: number
-  intricate: number
-}
+export type TransitionNone = 'none'
+export type TransitionRapid = 'rapid'
+export type TransitionQuick = 'quick'
+export type TransitionSimple = 'simple'
+export type TransitionModerate = 'moderate'
+export type TransitionComplex = 'complex'
+export type TransitionIntricate = 'intricate'
+
+export type Transitions =
+  | TransitionNone
+  | TransitionRapid
+  | TransitionQuick
+  | TransitionSimple
+  | TransitionModerate
+  | TransitionComplex
+  | TransitionIntricate
+
+export type TransitionRamp = Record<Transitions, number>